+++ data/src/main/java/com/fx_trading/lessons/data/api/lesson/ApiLesson.kt
@@ -15,7 +15,7 @@ data class ApiLesson(
     val language: String = "",
     val likes: Int = 0,
     val promo_image_url: String = "",
-    val questions_group: Long = 0,
+    val questions_group: Int = 0,
     val region: String = "",
     val sort_order: Int = -1,
     val speaker_name: String = "",
