+++ data/src/main/java/com/fx_trading/lessons/data/api/question_group/ApiQuestion.kt
@@ -3,7 +3,7 @@ package com.fx_trading.lessons.data.api.question_group
 data class ApiQuestion(
     val answers: List<ApiAnswer> = listOf(),
     val description: String = "",
-    val difficulty: Long = 0,
+    val difficulty: Int = 0,
     val active: Boolean = false,
     val optional_image_url: String = "",
     val title: String = ""
