+++ data/src/main/java/com/fx_trading/lessons/data/repositories/lessons/LessonsRemoteRepository.kt
@@ -4,6 +4,6 @@ import com.fx_trading.lessons.data.api.lesson.ApiLesson
 
 interface LessonsRemoteRepository {
     suspend  fun getRemoteLessons(): List<ApiLesson>
-    suspend fun getRemoteLessonByID(id: Long): ApiLesson?
+    suspend fun getRemoteLessonByID(id: Int): ApiLesson?
     suspend fun updateLesson(lesson: ApiLesson): Boolean
 }
