+++ data/src/main/java/com/fx_trading/lessons/data/repositories/lessons/LessonsRemoteRepositoryImpl.kt
@@ -39,7 +39,7 @@ class LessonsRemoteRepositoryImpl @Inject constructor(private val firebaseFirest
         }
     }
 
-    override suspend fun getRemoteLessonByID(id: Long): ApiLesson? {
+    override suspend fun getRemoteLessonByID(id: Int): ApiLesson? {
         try {
             val firebaseDocument = firebaseFirestore.collection("${documentPath}Lessons").document("$id").get().await()
             if (firebaseDocument != null && !firebaseDocument.data.isNullOrEmpty()) {
